delete in progress

+++ src/components/data-table.tsx
@@ -32,9 +32,11 @@ import {
     DropdownMenuTrigger,
   } from "@/components/ui/dropdown-menu"
 
+import { useConfirm } from "@/hooks/use-confirm"
 import { useState } from "react"
 import { Trash } from "lucide-react"
 import { toast } from "sonner"
+import { useToast } from "./ui/use-toast"
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
@@ -51,12 +53,17 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm(
+      "Are You Sure?",
+      "You are about to perform a bulk delete."
+    )
     const [sorting, setSorting] = useState<SortingState>([])
     const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>(
         []
     )
     const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({})
     const [rowSelection, setRowSelection] = useState({})
+    const { toast } = useToast();
 
   const table = useReactTable({
     data,
@@ -79,6 +86,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
         <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -95,10 +103,21 @@ export function DataTable<TData, TValue>({
                 size={'sm'}
                 variant={'outline'}
                 className="ml-auto font-normal test-xs text-red-600"
-                onClick={() => {
-                  toast.info("Deleting...")
-                  onDelete(table.getFilteredSelectedRowModel().rows)
-                  table.resetRowSelection()
+                onClick={async() => {
+                  const ok = await confirm();
+
+                  if(ok){
+                    toast({
+                      title: "Deleting",
+                      description: "Deleting the table(s)",
+                      variant: "destructive"
+                    })
+            
+                    onDelete(table.getFilteredSelectedRowModel().rows)
+                    table.resetRowSelection()
+                  }
+
+                  
                 }}
                 >
                     <Trash className="size-4 mr-2 bg-f"/>
